refactor ApiRequests.addMultiPeriodStream to only return data response without CSRF tokens
