+++ frontend/src/types/ModifyMultiPeriodStreamResponse.ts
@@ -1,9 +1,12 @@
 import { CsrfTokenCollection } from "./CsrfTokenCollection";
 import { MultiPeriodStream } from "./MultiPeriodStream";
 
-export interface ModifyMultiPeriodStreamJson {
-    csrfTokens: Partial<CsrfTokenCollection>;
+export interface ModifyMultiPeriodStreamResponse {
     errors: string[];
     success: boolean;
     model: MultiPeriodStream;
+}
+
+export interface ModifyMultiPeriodStreamJson extends ModifyMultiPeriodStreamResponse {
+    csrfTokens: Partial<CsrfTokenCollection>;
 }
